cloud_browse

cloud_browse() lets you navigate the folder tree interactively. This is useful for finding a file or folder path which can then be used in cloud_get() or cloud_put(). This function is only intended to be used interactively - you should not use this function programmatically.

When you call cloud_browse() you are given a list of files and folders (either at the top-level, or from the provided path). You provide the numeric number of the folder or file you wish to move to to continue. If you are not at the top level, you can select "../" to move up one folder. At any time you can select 0 to exit the interactive navigation.

If you select a folder, you are shown the files and folders within that folder. If you select a file, the full path for the file is shown (so that it can be passed to another function) and the interactive session is ended.

cloud_browse(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())

Arguments

path

The initial path to start the search. If not provided, the function starts at the top-level folder.

user

Your Cloudstor username.

password

Your Cloudstor password.

Value

the last file path